pyproject.toml
[tool.poetry]
name = "python-moje"
version = "0.1.0"
description = ""
authors = ["Anna Giszczak <anna.giszczak@cern.ch>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
rich = "^13.9.2"
argparse = "^1.4.0"
tqdm = "^4.66.5"
ascii_graph = "^1.5.1"
numpy = "2.0"
pillow = "^11.0.0"
numba = "^0.60.0"
bs4 = "^0.0.2"
selenium = "^4.27.1"
matplotlib = "^3.10.0"
scipy = "^1.14.1"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
